Introduce LastLevelCacheStore
LastLevelCacheStore is a version of CoreStore which has a stripe per
last level cache (as reported by folly::CacheLocality).

This is useful when heavy cache contention is noted in large systems,
but a stripe per core would be costly in terms of memory.

See folly::AccessSpreader for details on the allocation of cores to
stripes.

Change-Id: Ifba002fb9e44aa8d40c51addb5c99356e8e9c9db
Reviewed-on: http://review.couchbase.org/122659
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>