Refactor: don't use MemoryMappedFile for loadFile
Replace the use of MemoryMappedFile to load a file with
folly::readFile for unix and the Win32 API on windows
so that we can delete the MemoryMappedFile implementation.

Ideally we would use folly::readFile on all platforms, but
it cause a failure on windows as it don't allow for reading
a file someone else have open for writing.

Change-Id: Ice2cfeb5aff6504ed1aa8b7aec2759003619754a
Reviewed-on: http://review.couchbase.org/123751
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>