MB-38270: Add retry logic to cb::io::loadFile
ns_server may update the files (create tmp, delete old, rename tmp)
while memcached is trying to read the file. If memcached managed
to open the file before it was deleted everything is good (from
our point of view), but if we're unlucky we try to open the file
the file in the time gap between the delete and the rename.

The likelihood of hitting this issue is small, but we've seen it
the logs in a small number of MBs (most often as part of startup
that the audit file isn't present causing memcached to exit and
ns_server restarting memcached. This isn't a huge issue as memcached
hadn't started to warm up etc so the restart is quick, but we
could hit the same scenario elsewhere in our system. Instead
of adding the same logic all over the place its better to add
a parameter to loadFile to do the retry.

Replace the internal method to use folly::readFile so we can
kill the entire MemoryMappedFile class. Unfortunately
folly::readFile don't use the Windows API so not all return
codes are properly propagated so we need our own impl on
windows.

Change-Id: If467f8211e95fe0480a912bc1dd821a9b888d2bf
Reviewed-on: http://review.couchbase.org/123582
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>