Use boost::filesystem C++17 and folly
C++17 added support for std::filesystem which offers a platform
independed way to access the fileystem which we created our
own wrappers for in cb::io. Unfortunately none of the compilers
we use for linux and mac supports std::filesystem so we have
to stick with boost::filesystem for now.

Replace the implementation we've got for some of our helper
methods to use folly's wrappers.

Change-Id: I958e56d4931b47384c4bd62adeb95f997ff99d68
Reviewed-on: http://review.couchbase.org/123704
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>