AtomicDuration: Assert duration type is lock-free
Add a static assert that the Duration that AtomicDuration is
instantiated with is indeed a lock-free atomic.

The entire motivation for the custom AtomicDuration template class is
to provide an atomic, lock-free type with semantics the same as chrono
durations. If the user attempts (or doesn't realise) that their
duration type is not lock-free, then this class isn't suitable /
doesn't actually solve any problems which std::atomic<MyDuration>
doesn't already, and as such the user should be made aware of that.

Change-Id: Ia03b4218a6f17061f89810f38eeb1fa3a8475616
Reviewed-on: http://review.couchbase.org/c/platform/+/131101
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>