MB-42586: Initialise current in NonNegativeCounter
If we try to construct a NonNegativeCounter with a negative value
the exception could report 'current' as some unexpected value e.g.

  ThrowExceptionUnderflowPolicy current:140219260243264 arg:-1

In this case current should be ignored (but may require some
processing of logs/code to discover that). If we initialise current
the exception message is clearer e.g.

  ThrowExceptionUnderflowPolicy current:0 arg:-1

Change-Id: I16c5eac5da258e55a8a6c14c29e0984c16a75a4f
Reviewed-on: http://review.couchbase.org/c/platform/+/139902
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>