MB-38422: Set the jemalloc configuration on WIN32
On WIN32 the use of je_malloc_conf doesn't work, to set
the config we are now setting the JE_MALLOC_CONF
environment variable.

Todo this cb::ArenaMalloc now has a function that can be invoked
from executable (memcached/unit-tests) which "ensures" that
our configuration is set onto jemalloc - this is done on WIN32
by re-executing the process with an updated environment.

Change-Id: I0ab539b83723674ea92712f32e240df230c029dc
Reviewed-on: http://review.couchbase.org/c/platform/+/144631
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>