Cleanup: backtrace can use a C++ API
All usage of the backtrace methods are from C++ programs
so we don't need the C linkage anymore.

In addition drop the complexity of trying to build without
support for backtrace. If we ever want to support a new
platform we want to know at build time that the backtrace
support isn't there so we can fix it rather than building
a version and later detect that its not there.

Change-Id: I620813e09a7982357e4a62e6751c411a824b2dc9
Reviewed-on: http://review.couchbase.org/c/platform/+/145520
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>