cb::time2text: Round least significant component to nearest
Currently cb::time2text discards any remainder when converting a
number with a factional part - for example 10,999 ns is printed as "10
ms". This is misleading, if we are going to print values as whole
milliseconds, "11 ms" has a much lower error.

Modify time2text to round to the nearest integer when printing such
values.

Change-Id: Ieb40a0cd9add994e8c49f1eae4445a8562e38bd5
Reviewed-on: http://review.couchbase.org/c/platform/+/148169
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>