Extend ScopeTimer1 to allow for skipping timing
Call isEnabled() on the listener to avoid reading the system
clock. The motivation for this is to be able to use the
ScopedTimer for the client tracer spans in the core where
the client needs to enable via hello. If the client don't
enable the tracing we now won't read the system clock
to times when the class is used.

Change-Id: I93219cf6e6fc8cd7daf5785936561fc7dbcc2878
Reviewed-on: http://review.couchbase.org/c/platform/+/148991
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>