MB-45378: Convert 'platform' to static library
+Problem+

platform.a was previously just a small "stub" library which contained
the global new/delete replacements; and then itself depended on the
shared library platform_so.so, which contains all the actual platform
code (and links in all dependaneies such as Folly). As such, targets
statically linking 'platform' ended up linking to 'platform_so.so'
shared library which includes a copy of Folly.

Given we must end up with just one copy of Folly in memcached, we need
a platform library which is a standalone library, which contains all
platform code and doesn't introduce a dependancy on a shared library.

However, there's some additional complexity - we _also_ need to be
able to override the default functionality of cb_malloc / cb_free et
al symbols inside platform, to use our arena allocator for per-bucket
memory tracking. Previously this was done by defining cb_malloc... as
weak symbols in 'platform', then defining as strong symbols in static
library platform_cb_arena_alloc and then having targets which need
tracking linking to platform_cb_arena_alloc.

This approach falls down however if platform itself is a static
library and not shared - if the same symbol appears in two static
libraries the linker will simply take the first definiton it finds,
irrespective of weak/strong. Therfore if platform appears first then
we don't get the strong arena symbol from platform_cb_arena_alloc.

+Solution+

The main change here is to convert platform_so from a shared to static
library, renaming to 'platform' and removing the old stub 'platform'
library - this is mostly mechanical.

To address the static linking weak/strong symbol issue,
platform_cb_arena_alloc is converted from a static to object library -
given this passes the actual .o files to the linker individually we
are back to correct weak/strong behaviour - the strong symbols take
precedence.

Note that the export header (platform/visibility.h) temporarily
remains in this patch - while this is now unnecessary it's another
large change to remove it and hence will be done in a follow-up.

Change-Id: Ibf4e47c9db75d4084bf6f062156db9692b8132c6
Reviewed-on: http://review.couchbase.org/c/platform/+/151984
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>