MB-45378: Fix static init fiasco in RandomGenerator
RandomGenerator uses a static mutex and unique_ptr instance to manage
the its singleton. However, the order at which the mutex and unique_ptr
are initialised is undefined.

When converting platform from a shared to static object, under VS2017
the order of construction of the two is reversed - i.e. the mutex is
*not* created first and hence we see a segfault trying to lock the
mutex.

Note this is a problem in general, unrelated to the static library
change - we've basically "got lucky" previously.

Fix by using a C++11 "magic static" which handles all this automatically
in a thread-safe way.

Change-Id: I75bee4d5be571ce650c758072f6aea59ee1c5269
Reviewed-on: http://review.couchbase.org/c/platform/+/152006
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>