MB-46017: Add ProcessMonitor
Add a process monitor which allows to either start a child process
and monitor the life of that process, or monitor the existence
of another process.

The class started out as a copy of the ParentMonitor used in
memcached, but added support for creating the child process
to be used in testapp, cluster_test etc.

Change-Id: I1106e7d240525183527105f70023c7088224d287
Reviewed-on: http://review.couchbase.org/c/platform/+/153320
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>