MB-46257: Close filehandle if bad_alloc is thrown
cb::io::loadfile could leak a handle to the file if std::bad_alloc
was thrown while resizing the buffer to hold the content of the
file. Leaking filehandles on Windows is even more severe than on
Unix platforms as it blocks other proecsses from removing the file
until the process terminates.

Change-Id: I55237d9e454a4dfc6de2698f991edbe1d59673ff
Reviewed-on: http://review.couchbase.org/c/platform/+/153423
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>