MB-46988: Fix CLion project indexing with CB_PCH
When using precompiled headers and CLion; when it attempts to
preprocess each source file for indexing / analysis it fails with an
error about being unable to find <gtest/gtest.h>.

The issue here seems to be related to how CLion invokes the
preprocessor - it needs the include path for all headers inside
precompiled_headers.h; even though that has already been precompiled.

If we make the PCH target include paths public (i.e. so they are
inherited by consumers of the PCH target), then this addresses the
issue.

One minor fly in this ointment is that ranlib now gives warnings about
precompiled_headers.cc.o having no symbols - e.g

    ranlib: file: kv_engine/libmcbp.a(precompiled_headers.cc.o) has no symbols

This seems benigh but is a bit chatty, so suppress it by defining a
dummy symbol in precompiled_headers.cc.

Change-Id: I122c1b0693251b5d0cbb8de9010e01b38faabc43
Reviewed-on: http://review.couchbase.org/c/platform/+/156121
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>