CBD-4127: Only configure jemalloc profiling for Linux
jemalloc only supports profiling on Linux; however prior to cbdep
jemalloc-5.2.1-cb6 we did configure jemalloc with profiling on for all
platforms. That was somewhwat confusing / broken, as while you could
attempt to turn it on on macOS, it didn't actually do anything.

Subsequent to jemalloc-5.2.1-cb6 profiling is (correctly) not present
in our jemalloc library, so if we attempt to enable it we see the
following error:

    <jemalloc>: Invalid conf pair: prof:true
    <jemalloc>: Invalid conf pair: prof_active:false

Fix this by not attempting to configure profiling on non-Linux
platforms.

Change-Id: Ia4ce00f625dad7e9f3025fc7d3299331a237ba07
Reviewed-on: http://review.couchbase.org/c/platform/+/157148
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>