MB-47308: Add MemoryDomain to ArenaMalloc
This commit extends the ArenaMalloc API to support two memory domains.
A domain is a coarse sub-division of the arena's memory usage.

The user of ArenaMalloc can set the domain, for example the secondary
domain could be set before calling into a library.

Memory usage can then be queried by total (sum of all domains) or each
domain queried individually.

The API of ArenaMalloc is extended as follows:

A version of switchToClient is available where the caller can select
the domain, the existing switchToClient defaults to
MemoryDomain::Primary.

setDomain is added which allows the caller to select the domain that
will apply to all subsequent memory alloc/dealloc calls.

getPreciseAllocated and getEstimatedAllocated now have functions
which take a domain and return the memory usage for that domain.

Change-Id: I42c955317670f822e1be080102053c97973fde95
Reviewed-on: http://review.couchbase.org/c/platform/+/157254
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>