Add cb::net::initialize to initialize sockets
Rename cb_initialize_sockets now that we're C++ only and already
have a namespace with all of the socket functions.

By making it a function on all platforms we no longer get a warning
in CLion for unused statements. I was initially thinking that we
could modify the code to throw an exception, but then again all
places where we use sockets we cannot continue if we fail to
initialize sockets so we could might as well just terminate
the program without having to distribute the logic to do so.

Change-Id: I376535c8d27195dd1fbdf33db07f3b11c28d6ef9
Reviewed-on: http://review.couchbase.org/c/platform/+/157939
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>