Reduce cost of including platform/exceptions.h
platform/exceptions.h is an expensive header to include - it is
included from many places and the per-inclusion cost is high - ~350ms
- from clang-build-analyser building `kv_engine_everything` target:

    42500 ms: ../platform/include/platform/exceptions.h (included 121 times, avg 351 ms)

The main reason for the high cost is that it includes template
functions whose definition is inline, and definition requires
including a number of costly Boost stacktrace/exception headers.

Given we only instantiate those template functions with a small number
of types, move the template definitions to a new exceptions.cc file,
providing explicit template instantiations for the types we need.

This reduces the overall build time by 75s (or 3%):

From:

    Compilation (678 times):
      Parsing (frontend):         1743.2 s
      Codegen & opts (backend):    536.6 s

To:

    Compilation (677 times):
      Parsing (frontend):         1684.8 s
      Codegen & opts (backend):    520.0 s

And the header itself now cost 6ms to include:

    761 ms: ../platform/include/platform/exceptions.h (included 120 times, avg 6 ms)

Change-Id: I172d2a5d51b2f259997a01baccb8f9f13e4d6355
Reviewed-on: http://review.couchbase.org/c/platform/+/160306
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>