MB-48688: Introduce estimateClockOverhead
Add a new function to timeutils which estimates the overhead of
measuring the current time with the given clock - the time taken to
perform Clock::now() over a number of samples.

This will be used in memcached to print the overhead of the main
clocks it uses at startup, and also to report the overheads via
stats. This can be useful when debugging OS / environments where
reading the clock is costly - for example some virtualised
environments.

Change-Id: I274a691a26cab3ae38481c33cc0566e6c244062d
Reviewed-on: http://review.couchbase.org/c/platform/+/162649
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>