MB-49867: Fix problem with mixing V1 and V2
On Ubuntu 20.04 (when running on the bare metal) it has
a cgroup V2 mounted but the memory,cpu and cpuacct controller
is mounted via V1.

Processes started would end up in cgroup V1 for memory,cpu,cpuacct
_AND_ in the unified V2 (but not listed as enabled).

Previously we preferred V2 and fell back on V1, but we should
search V1 first and if we find any of the ones we use we should
use V1 and fall back to V2 if none was found.

Change-Id: I4bbee206325f3788a7f907954e0bc0eb252c5d96
Reviewed-on: https://review.couchbase.org/c/platform/+/166842
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>