MB-49983: MemoryTrackingAllocator: support sized deallocation
MemoryTrackingAllocator is a proxy allocator which tracks the total
amount of bytes allocated, delegating to global new/delete to actually
provide memory.

However, the deallocate() method does not take advantage of the size
information passed into it - it just calls 'delete(void*)' and not the
sized 'delete(void*, size_t)'. This has a significant impact on
performance; particulary when using our memory-tracking global
new/delete - as they need to determine the size of the allocation
which is being freed to account for it. When vanilla (non-sized)
delete(void*) is called, we must use an expensive call to je_sallocx
to lookup the size from the pointer. That is not the case if
sized-delete is used, as the allocation size can be cheaply calculated
from the requested size.

Modify MemoryTrackingAllocator to use std::allocator<T> for allocation
and deallocation; as that alreasy handles sized-delete (and aligned
also).

Change-Id: I1d87ce53ec3b2a6abf4b03a6511d618d149b2c92
Reviewed-on: https://review.couchbase.org/c/platform/+/167243
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>