MB-46378: AtomicNonNegativeCounter [1/2]: Add AtomicNonNegativeCounter
NonNegativeCounter is currently implemented using std::atomic<> around
the templated value type, and suitable use of compare-exchange to
enusre that NonNegativeCounter itself is atomic - it can be safely
read and/or written from multiple threads much like std::atomic.

However, for cases where we do not need this atomicity (where a
counter is not accessed concurrently by multiple threads) we are
paying the additional cost of the atomic operations. On x86 with its
strong memory ordering this cost is small, but for other architectures
such as AArch64 which has a weaker memory ordering by default the cost
is higher - sufficiently so that there is value in having both atomic
and non-atomic versions of NonNegativeCounter and using the non-atomic
version where possible.

This patch introduces an explicitly atomic version of
NonNegativeCounter: AtomicNonNegativeCounter. This is essentially a
copy-paste of the current NonNegativeCounter; to allow us to introduce
the new type without changing the semantics of the existing
NonNegativeCounter while code is migrated over.

Once migration is complete, NonNegativeCounter will be changed to not
use std::atomic and remove any unnecessary code (e.g. compare-exchange
retry loops).

(Note this approach is consistent with the Monotonic class templates,
where we have both Montonic and AtomicMonotonic - the former being
non-atomic).

Change-Id: I3e6843c4eace98534ce039fb840c5068a5b5b781
Reviewed-on: https://review.couchbase.org/c/platform/+/167325
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>