MB-46360: maximizeFileDescriptors: Fix overflow on macOS / AArch64
When running IoTest.maximizeFileDescriptors on macos/arm64, the test
hangs forever.

The issue is during the last test case where we call with a limit of
uint64_t::max(), the 'avg' calculation overflows (min + max), and on
AArch64 this ends up giving the same result for 'avg' as before and
the while loop never exits.

Fix by avoiding adding the min + max - using the same algorithm as std::midpoint from C++20.

Change-Id: I94d0f4d5a265f8c9e2871b020141d730c0a11ce2
Reviewed-on: https://review.couchbase.org/c/platform/+/170643
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>