MB-50988: Introduce cb::AwaitableSemaphore
cb::Semaphore is a simple non-blocking semaphore, and is used to limit
concurrency for tasks like PagingVisitor.

However, if a token cannot be acquired (i.e., they are all held by
other actors), cb::Semaphore does not provide any mechanism to notify
callers when a token becomes available.

For a "standard" semaphore with blocking like std::counting_semaphore,
callers blocked waiting for a token can be woken when a token has been
released.

Our task/thread pool model would not work well with blocking, but
notification would be useful.

To support this, introduce AwaitableSemaphore, which can track Waiters,
and can signal them once tokens have been made available.

In the context of tasks, this could be used to wake a sleeping task.

Change-Id: I9403f97f51f4eba1e497b1f74cc6bb19743c5261
Reviewed-on: https://review.couchbase.org/c/platform/+/171406
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>