MB-51556: Modify monotonic.h to use template parm based labels
Modify our Monotonic templates so that the policies now take a label as
an optional template parameter, rather than being dynamically assigned
at run time. Requiring us to store a sizeof(std::string) + heap space
for each instance of the var. This use non-insignificant amounts of
memory for variables in classes where large number of objects from.

By making the label a template we can use a constexpr static to store
the label which will be used by all instances a class's member variable
when throwing. Meaning our memory usage will scale O(1) rather than
O(n).

In this patch we also add the ability to pass a ref to a Labeler object
that stores a lambda to process the static label, allowing the use to
add extra information at run time to the string.

Change-Id: I61332b72fee6a5a87fdec23c3b1b3c5b69e37347
Reviewed-on: https://review.couchbase.org/c/platform/+/171269
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>