MB-50988: Fix AwaitableSemaphore edge case
AwaitableSemaphore maintains a unique queue of waiters which have
requested notification once a token becomes available.

However, having the queue just be unique is insufficient if waiters can
be woken by anything other than the semaphore.

In that case, if a waiter successfully acquires a token, it must ensure
it is not _also_ still in the queue.

If it does not check this, a later release() may notify the same
waiter, which already has a token - this may lead to a "lost"
notification.

In the "normal" path, if the semaphore woke the waiter, it will not be
in the queue, which can be cheaply checked.

It is expected that it is relatively uncommon for a waiter to be woken
by anything else, and thus the extra cost of manipulating the queue in
that path is not a large overhead.

Change-Id: I363525a5397fecaadee1cd6334233eeede2a731d
Reviewed-on: https://review.couchbase.org/c/platform/+/171998
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>