MB-51608: Use automatic jemalloc tcache selection
This commit disables per bucket tcaches in favour of allowing jemalloc
to choose the tcache for us. Effectively this means all buckets are now
sharing a tcache. The result of this is possible some "drift" in the
ArenaMalloc stats as a bucket may effectively allocate memory out of
the tcache which is memory from another bucket's arena.

This now means we won't hit the upper bound of 4093 tcaches when we
have many buckets and many threads.

Change-Id: I5bed7b44b3cc8db52cf222d1edd61e0b37291a92
Reviewed-on: https://review.couchbase.org/c/platform/+/172970
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>