MB-51912: Workaround msvc v19.16 compiler bug
Problem:
MSVC v19.16 and below contain a compiler bug that means we can't use
a const char* template pram inside a template at run time. This
affects monotonic.h, as we use the Name template argument when throwing
an exception if the assignment is not monotonic.

Fix:
For now add a '#if defined(_MSC_VER) && (_MSC_VER <= 1916)' to not use
the Name template pram if the compiler is affected by this bug.

In the long run we should remove this if statement when we've upgraded
to msvc v19.20 or newer (MB-43718).

Change-Id: I9109cbafbf95c89050823cbe7898dd95d684fa21
Reviewed-on: https://review.couchbase.org/c/platform/+/174047
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>