CBD-6348: [BP] Changes required for folly v2022.02.14.00
[[ Linked patch with tlm folly upgrade to v2022.02.14.00 ]]

Folly v2022.02.14.00 has removed the method getUniquePtr(), replaced
with as_lock().

Unfortunately our previous version of folly (v2020.09.07.00) did not
yet have the as_lock() function, so we cannot update platform in
advance of the folly update - we must change the method in lock-step
with the upgrade.

Change-Id: I13e45b8bb9aa59cdc0dcb5a754b1674deabed0d8
Reviewed-on: https://review.couchbase.org/c/platform/+/234763
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Well-Formed: Restriction Checker