Use folly::setThreadName()
Folly claims to be able to set the thread names on all of our
platforms so we should let it do so. Unfortunately the win32
implementation in folly seems to fail to return the thread
name, but it may be because of the thread name API was
added in Windows 10 build 1607. Who knows. Anyway the
thread names in win32 isn't available in taskmanager, but
should appear in the debugger)

Change-Id: I6b2d34b481df3628745417bb701181ed7019ac95
Reviewed-on: https://review.couchbase.org/c/platform/+/175204
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>