MB-54106 : Build a libplatform variant that doesn't override allocator symbols.
Problem: Whenever a major GC run is triggered by the newly upgraded
V8(v10.7.21), there is a call to system malloc followed by
malloc_usable_size. This always results in a crash if the embedder
of V8 also links to the "platform" library.
This is because "platform" replaces the implementation of
malloc_usable_size to its jemalloc equivalent.

Solution: If any V8 embedding service also happens to link to platform,
it should instead link to "platform_without_global_new_replacement".
This target is same as "platform" but without replacing symbols
new / delete and malloc_usable_size

Change-Id: I113872301bbaf622bd07c3ce3e1afaab77246ec1
Reviewed-on: https://review.couchbase.org/c/platform/+/181084
Well-Formed: Restriction Checker
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>