MB-54297: Don't override operator new/delete for ASan build
When building with AddressSanitizer it already overrides operator
new/delete so it can track heap usage. While it is currently benign to
include global_new_replacement.cc in platform with the ASan build as
ASan "wins", a future change which moves global_new_replacement.cc
into a different object library results in link-time errors as the
linker finds duplicate symbols.

As such, remove when building with ASan.

Change-Id: Ib36c2c4258e5468fba6c982f7bd46e8ee7b99d61
Reviewed-on: https://review.couchbase.org/c/platform/+/182142
Reviewed-by: James H <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>