MB-54810: HdrHistogram: track samples which overflow bucket range
Add an 'overflowed' counter to HdrHistogram which tracks the count of
samples added to the object, but which exceed the maximum trackable
value of the underlying hdr_histogram. Include this field in the JSON
serialisation of the HdrHistogram. Also include a 'max_trackable'
field which specifies above what value the overflow count refers to.

Note that as these are tracked outside hdr_histogram, they are not
included in hdr_histograms' own percentile calculation / iterators
etc.

Howver the overflow count can at least be reported by of the histogram
in a suitable way, so they are flagged to the user.

Change-Id: I64f4b7354affcf6a0f9a74a2885e8a812d9dfe45
Reviewed-on: https://review.couchbase.org/c/platform/+/186868
Well-Formed: Restriction Checker
Reviewed-by: James H <james.harrison@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>