MB-54810: HdrHistogram: track overflowed samples sum
Expand the information we track about samples which overflow the
trackable range of HdrHistogram, to include the sum of these samples.

This information can be used by consumers of HdrHistogram to calculate
a mean of the values recorded which includes the samples which
overflowed. For example, Prometheus requires a sum and count field for
its histograms (it can then derive the mean itself). With this change
the overflowed sum can be included in the 'sum' field.

Change-Id: I06cf8ac4bd7bed6d416890d50e1fa96c55bee269
Reviewed-on: https://review.couchbase.org/c/platform/+/187139
Well-Formed: Restriction Checker
Reviewed-by: James H <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>