Add cb::getopt::CommandLineOptionsParser
Add a callback driven C++ version on top of getopt_long
which hides builing up the option array and the format
string for the short options. It also takes care
of looping over the various arguments passed to the
program and finally build up a vector containing
all of the unparsed arguments.

Change-Id: I7999254f23b1bfe74d1157ca2896de5115cf4492
Reviewed-on: https://review.couchbase.org/c/platform/+/188593
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>