Allow for EBO within [Atomic]Monotonic
By privately inheriting from the LabelPolicy, as opposed to having it
declared as a member, we allow compilers to use Empty Base Optimisation.

This avoids wasting space for policies which are empty classes, such as
the IgnorePolicy, and reduces the size of Monotonic<int64_t> to 8 bytes,
down from 16 (where 1 byte was previously used for the labeler, and 7
bytes was padding at the end of the object).

Change-Id: I515170d0f7b9218eb6b16e13fbcdbcc3cb06655e
Reviewed-on: https://review.couchbase.org/c/platform/+/190909
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>