MB-58644: JEArenaMalloc: validate alloc/free have matched client & domain
Add a debugging mode to JEArenaMalloc which validates that memory is
always deallocated against the same client (bucket) index and domain
as it was allocated. This helps to ensure that mem_used and related
stats in ObjectRegistry are correct - we account deallocations to the
currently selected client; but if that is *not* the client which
originally allocated the memory, mis-accounting would occur.

Note that this check requires that jemalloc tcaches are disabled,
otherwise the areans.lookup check is not valid.

The check is done by taking advantage of two jemalloc features -
multiple arenas, and arena.lookup to find which arena an allocation
belongs to.

Instead of creating one arena per client (bucket), we create mutliple,
one for each Domain (Primary - for ep-engine, and Secondary - for
Magma). When switching between different Domains on the same bucket,
we now switch between the two arenas assigned to that Bucket, so
allocations made against say the Primary domain go into a specific
arena.

Later, when memory is freed - which _should_ occur from the same
client & domain as it was allocated - we ask jemalloc which arena the
to-be-freed allocation belongs to, and verify that is the same as the
arena associated with the currently selected bucket+domain. On a
mismatch, we print details of the mismatch including backtrace, and
abort().

The checks require the use of JeArenaSimpleTracker - which is enabled
for Debug builds (NDEBUG not defined) only.

Change-Id: I2f3b2d627be753d6e89a3aa68668ba3013feb839
Reviewed-on: https://review.couchbase.org/c/platform/+/198384
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>