MB-58714: Introduce NoArenaGuard
Add 'NoArenaGuard' - a helper class which allows memory tracking to be
temporarily switched away from current client.

This is useful when calling into code which is not client (aka Bucket)
specific, and which performs memory allocations - we do not want to
account these (de)allocations to a specific Bucket; instead they
should be accounted to the "NoClient" / "NoArena". Creating a
NoArenaGuard makes this very easy - the caller doesn't need to know
which particular client is currently running.

Note: Linked with corresponding kv_engine change (return value of
switch{From,To}Client changing requires changes in kv_engine).

(Note to the wider scope of the system - NoArenaGuard is similar to
NonBucketAllocationGuard, however it does not depend on any ep-engine
code (unlke NonBucketAllocationGuard). This is important because we
have code which is independent of ep-engine (e.g memcached_logger),
and hence cannot use ep-engine code in it - but we _do_ want to
disable memory tracking at certain times.

Change-Id: I37b7ca4d7391ed0ab296980fa49c429b437fe2cf
Reviewed-on: https://review.couchbase.org/c/platform/+/197470
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>