MB-58714: Make SystemArenaMalloc's ClientAndDomain public
To add support for a 'NoArenaGuard' helper class (which allows memory
tracking to be temporarily switched away from current client) we need
to expose a ClientHandle from each ArenaMalloc implementation when
switching from and back to a client.

For SystemArenaMalloc this is the ClientAndDomain class, however this
is currently private to the JEArenaMalloc implementation. Move it to
the public API in preparation for using it in the NoArenaGuard impl.

Change-Id: Id71a6726f9ca4bafbfb43c7bd4fbf27f9a74cf2b
Reviewed-on: https://review.couchbase.org/c/platform/+/197469
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>