MB-58714: Make JEArenaMalloc's CurrentClient public
To add support for a 'NoArenaGuard' helper class (which allows memory
tracking to be temporarily switched away from current client) we need
to expose a ClientHandle from each ArenaMalloc implementation when
switching from and back to a client.

For JeArenaMalloc this is the CurrentClient class, however this is
currently private to the JEArenaMalloc implementation. Move it to the
public API in preparation for using it in the NoArenaGuard impl.

Change-Id: I2469f91ed91c822c728bc66c9a3f612a560b206f
Reviewed-on: https://review.couchbase.org/c/platform/+/197468
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>