MB-53572: Add ::to_string(time_point)
Add a new method to format a time point into a string.
This method currently use spdlog, which may be replaced
with fmt directly once we upgrade fmt.

It is added as a new method in platform because:

a) It is used from multiple repositories
b) Allow for writing unit tests that the new implementation
   produce the same result as the old implementation

Change-Id: I29854e8df39ce1faeba8ec5d352a005a92ea38be
Reviewed-on: https://review.couchbase.org/c/platform/+/198190
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>