[BP] MB-61464: Add ::to_string(time_point)
BP of MB-53572 to 7.2.6.

Add a new method to format a time point into a string.
This method currently use spdlog, which may be replaced
with fmt directly once we upgrade fmt.

It is added as a new method in platform because:

a) It is used from multiple repositories
b) Allow for writing unit tests that the new implementation
   produce the same result as the old implementation

Change-Id: I29854e8df39ce1faeba8ec5d352a005a92ea38be
Reviewed-on: https://review.couchbase.org/c/platform/+/212638
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>