MB-56926: Add regression tests for oversize allocations
jemalloc has a feature for allocating "oversized" allocations in a
separate hidden arena. This is enabled by default and controlled by the
opt.oversize_threshold which defaults to 8 MiB.

This has the effect that any "oversize" allocations are not counted in
the expected arena allocation stats.

The oversized allocation feature is disabled for any requests which
explicitly specify the arena to use using MALLOCX_ARENA.

We used to not specify MALLOCX_ARENA for allocations outside of any
bucket. Later we made a change which always specifies MALLOCX_ARENA,
essentially disabling the "oversized" allocation arena, but we never
added regression tests for it, as the change was accidental.

Change-Id: If309df966a040c0c4f447b36306bfa138b6ce300
Reviewed-on: https://review.couchbase.org/c/platform/+/205099
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>