Add cb::net::set_socket_blocking
There is a utility function to set a socket into non-blocking mode
(on top of libevent), but in some cases we want to flip the
socket back to blocking mode.

This code is copied from kv-engine's client_connection.cc
used in the unit tests as part of releasing the socket. Given
that I have need for it elsewhere it is better to refactor it
to a common method for reuse in both locations.

Change-Id: Ibb9d7d3ef56aa9051125802efa9a8ae9679c756a
Reviewed-on: https://review.couchbase.org/c/platform/+/205570
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>