MB-55290: Add cb::NonBlockingSharedMutex
...as an alternative to std::shared_mutex, with the difference that this
implementation is not allowed to fail spuriously on try_lock*.

The use-case is when we would have to wait for a long-running task to
release a lock before our acquire succeeds, and yielding the user-space
task is preferable to blocking. We would still like to be certain that
an other object holds a lock.

Change-Id: Iacd2667467b1e88d7c910e1fce3a9417999af873
Reviewed-on: https://review.couchbase.org/c/platform/+/206201
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>