CBD-5660: Add missing headers for malloc_usable_size
We used to get malloc_usable_size included "for free" with the
previous version of folly via <folly/portability/GTest.h>, but
they have shuffled headers around and hence we now must include
it explicitly ourselves.

This change also adds some headers to src/backtrace.cc which
are required for building with msvc2022.

Change-Id: Id64b57c80bdb5ecb2e4520078f872e05bae9e086
Reviewed-on: https://review.couchbase.org/c/platform/+/201829
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>