MB-60936: Add cb::Unshared for fast contended counters
This is the same idea behind the je_arena_corelocal_tracker.h, but
refactored into a reusable template.

The template can be used to pack several counters together, as we have
done for the domain allocated memory, or it can be used with a single
counter (although this wastes a lot of space).

With this template in place, we can take replace all of the allocation
counting logic from je_arena_corelocal_tracker.h with a simple:

  struct ClientData {
    Unshared<MemoryDomain> allocated;
  };

The cb::Unshared with pick up the MemoryDomain::Count as the number of
counters, it will maintain core-local deltas, estimates and a sum of the
estimates (total allocation size of all domains).

The methods in the tracker then just call into cb::Unshared.

  getEstimatedAllocated() -> allocated.getEsimateSum();
  getEstimateAllocated(Domain domain) -> allocated.getEsimate(domain);

Named cb::Unshared as I saw a library called mcounter which does
something similar and uses that word.

Change-Id: I82411ba0f3bc2cdc2837f9b75f9ca987db47e04d
Reviewed-on: https://review.couchbase.org/c/platform/+/207196
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>